refactor: move react contexts out of components directory
