+++ site/src/components/Navbar.tsx
@@ -18,8 +18,8 @@ import { Link } from "./Link";
 import { BlockProtocolLogoIcon } from "./SvgIcon/BlockProtocolLogoIcon";
 import { BoltIcon } from "./SvgIcon/BoltIcon";
 import { HOME_PAGE_HEADER_HEIGHT } from "../pages/index.page";
-import SiteMapContext from "./context/SiteMapContext";
-import UserContext from "./context/UserContext";
+import SiteMapContext from "../context/SiteMapContext";
+import UserContext from "../context/UserContext";
 import { AccountDropdown } from "./Navbar/AccountDropdown";
 import { MobileNavItems } from "./Navbar/MobileNavItems";
 import { itemIsPage, NAVBAR_LINK_ICONS } from "./Navbar/util";
